Rename createClient function
